refactor(vselect): convert tests to vue-test-utils
